chat_app/requirements/base.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
environ==1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2022.2
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
